languageconverter/requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
python-docx==1.1.2
PyPDF2==3.0.1
deep-translator==1.11.4
reportlab==4.2.2
langdetect==1.0.9
